04 auth

+++ model/index.js
@@ -1,28 +1,45 @@
 const Contact = require("./schemas/contact");
 
-const listContacts = async () => {
-  const data = await Contact.find();
-  return data;
+const listContacts = async (userId, { limit = 5, page = 1, sortBy, sortByDesc }) => {
+  const { docs: contacts, totalDocs: total } = await Contact.paginate(
+    { owner: userId },
+    {
+      limit,
+      page,
+      sort: {
+        ...(sortBy ? { [`${sortBy}`]: 1 } : {}),
+        ...(sortByDesc ? { [`${sortByDesc}`]: -1 } : {}),
+      },
+      populate: {
+        path: "owner",
+        select: "email subscription -_id"
+      }
+    }
+  );
+  return { contacts, total, page: Number(page), limit: Number(limit) };
 };
 
-const getContactById = async (contactId) => {
-  const contactById = await Contact.findOne({ _id: contactId });
+const getContactById = async (userId, contactId) => {
+  const contactById = await Contact.findOne({ _id: contactId, owner: userId }).populate({
+    path: "owner",
+    select: "email subscription -_id"
+  });
   return contactById;
 };
 
-const removeContact = async (contactId) => {
-  const contactToRemove = Contact.findByIdAndRemove({ _id: contactId });
+const removeContact = async (userId, contactId) => {
+  const contactToRemove = Contact.findByIdAndRemove({ _id: contactId, owner: userId});
   return contactToRemove;
 };
 
-const addContact = async (body) => {
-  const newContact = await Contact.create(body);
+const addContact = async (body, userId) => {
+  const newContact = await Contact.create({ ...body, owner: userId });
   return newContact;
 };
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (userId, contactId, body) => {
   const updatedContact = await Contact.findByIdAndUpdate(
-    { _id: contactId },
+    { _id: contactId, owner: userId },
     body,
     { new: true }
   );
